+++ examples/react/src/hooks/marker-io.ts
